create new mod_state file for custom expectations
